Include test routes; rename "plantAPI"
