+++ garden-harvest/src/components/Plants/PlantDetails.js
@@ -1,5 +1,5 @@
 import React, {useEffect, useState} from 'react'
-import {fetchPlantDetails} from '../../api/PlantAPI'
+import {fetchPlantDetails} from '../../api/plantAPI'
 
 function PlantDetail(props) {
