+++ garden-harvest/src/components/Routes.js
@@ -1,20 +1,20 @@
 import React from "react";
 import {Route, Switch} from "react-router-dom";
-import Home from "../containers/Home";
-import CreateAccount from "../containers/CreateAccount";
-import Dashboard from "../containers/Dashboard";
+import Calendar from "./Calendar"
+import AllPlants from './Plants/AllPlants'
+import PlantDetails from './Plants/PlantDetails'
 
 export default function Routes() {
   return (
     <Switch>
-      <Route exact path="/">
-        <Home/>
+      <Route exact path="/TestingCalendar">
+        <Calendar/>
       </Route>
-      <Route exact path="/CreateAccount">
-        <CreateAccount/>
+      <Route exact path='/plants/'>
+        <AllPlants/>
       </Route>
-      <Route exact path="/Dashboard">
-        <Dashboard/>
+      <Route exact path='/plants/:plant_id/'>
+        <PlantDetails/>
       </Route>
     </Switch>
   );
